Working version. "import" and "extend" don't work
This raises a potential question about how this *should*
work. Does it make sense for there to be multiple templating systems,
and thus, require those to interact. Otherwise this becomes easier,
but I need to duplicate the standard Revel messages for Amber (and for
every other template).

One potential system would specify a template for user views, then
simply use the defaults for errors.